Refactor List components name specifically
